+++ src/g_game.h
@@ -124,3 +124,4 @@ extern gameaction_t loadaction;
 extern bool         newpars;
 extern int          pars[10][10];
 extern int          cpars[100];
+extern bool         resetinventory;